pyproject.toml
[project]
name = "clearIP"
version = "0.0.0"
description = "IP 地址信息查询"
authors = [{ name = "QING", email = "qlear@qq.com" }]
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["pytest>=8.3.5", "ruff>=0.11.0"]


[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
